requirements.txt
requests
pytest
faker
python-dotenv
pydantic
allure-pytest
pytest-check
pytest-playwright
pytest-mock
pymysql
sqlalchemy